refactor: finish modules
