+++ apps/frontend/src/components/ui/servers/notice/NoticeDashboardItem.vue
@@ -3,11 +3,8 @@ import dayjs from "dayjs";
 import { ButtonStyled, commonMessages, CopyCode, ServerNotice, TagItem } from "@modrinth/ui";
 import { EditIcon, SettingsIcon, TrashIcon } from "@modrinth/assets";
 import type { ServerNotice as ServerNoticeType } from "@modrinth/utils";
-import { useRelativeTime } from "@modrinth/ui";
-import {
-  getDismissableMetadata,
-  NOTICE_LEVELS,
-} from "@modrinth/ui/src/utils/notices.ts";
+import { useRelativeTime, getDismissableMetadata,
+  NOTICE_LEVELS, } from "@modrinth/ui";
 import { useVIntl } from "@vintl/vintl";
 
 const { formatMessage } = useVIntl();
