+++ packages/utils/servers/errors/index.ts
@@ -1,2 +1,3 @@
 export * from "./pyro-server-error";
+export * from './pyro-fetch-error';
 export * from "./server-error"
