+++ packages/utils/servers/errors/pyro-fetch-error.ts
@@ -1,0 +1,10 @@
+export class PyroFetchError extends Error {
+  constructor(
+    message: string,
+    public statusCode?: number,
+    public originalError?: Error,
+  ) {
+    super(message);
+    this.name = "PyroFetchError";
+  }
+}
